[contacts] Improve compression algorithm and provide details of compressed groups. Contributes to JB#45907
This simplifies the compression algorithm and creates a much more
consistent/predictable result. The number of markers is calculated
before compression begins instead of starting in the middle and
working outwards and looping back to the centre when more characters
need to be compressed. It also avoids imbalanced compressions where
significantly more characters are compressed in the first vs the
second half of the list.

Also SeasideStringListCompressor::compress() now provides details of
the compressed contents via a CompressedContent argument. These
details are passed on in SeasideDisplayLabelGroupModel in the
CompressedContentRole.