Use QTEST_MAIN macro instead of QTEST_APPLESS_MAIN
qtcontacts-sqlite engine now attempts to store a pointer to itself
into a dynamic QObject property of the application, to allow extension
aware code to retrieve it and operate with it (e.g. to connect to
extension-specific signals).

This will crash if there is no QCoreApplication instantiated e.g. in
the unit test, so this commit ensures that an application instance
does exist, by calling QTEST_MAIN instead of QTEST_APPLESS_MAIN.