[nemo-qml-plugin-contacts] Use multi-char compression marker. Fixes JB#46544
The ListCompressor compresses multiple entries into groups and uses the
compression marker to signify this. Previously the compression marker
was set as a fullstop '.', however, this conflicts with any individual
database entries that start with a fullstop. This results in entries
keyed with a '.' in the SeasideDisplayLabelGroup::m_groupIndices hash
map that represent both groups and entries.

Since groups are mostly represented by a single character (the first
character of the entries in the group), changing the compression marker
to a multi-character string should avoid such conflicts from occurring.
Any future label group plugins must avoid using ".." as a group label to
prevent the problem reoccuring.

The unit tests have been updated to reflect this change. Some more tests
were also added to check a slightly broader range of character entries.

This commit also avoids an invalid return reference if the display label
value is missing. If an entry is missing from the
SeasideDisplayLabelGroupModel::m_compressedContent QMap, the value()
request returns an invalid const reference. Providing a default value
that lives in the same scope avoids this.

It's not clear that this last change is particularly valuable, since if
the value doesn't exist it causes problems elsewhere. In theory the
value should always exist. It just makes this particular method more
robust.