[nemo-qml-plugin-contacts] Implement sorted search filtering. Contributes to JB#47217
We cannot perform the filtering via a backend request, because
we do a variety of locale-specific filtering things in-memory
(e.g. splitting word boundaries, ignoring diacritics, etc).

In the future, once we daemonise the contacts API, we should
move that functionality into the daemon, but for now we have
to do it in-memory in the SeasideFilteredModel.