[nemo-qml-plugin-contacts] Add special case handling for matching nicknames. Contributes to JB#49174
Nickname details are exposed to QML as a list of objects (which
include a variety of information including provenance/sync target)
rather than as a list of strings.

This commit ensures that we transform the nicknames to a list of
strings when exposing that data via the match data role.