[contacts] Use cache display label where possible. JB#50658
In the past, display names were not typically set in QContactDisplayLabel
so SeasidePerson had to calculate a display label. This is no longer
the case, so the QContactDisplayLabel value should be preferred.

In cases where the cache item does not exist (e.g. a temporary
contact not saved to db), SeasidePerson still needs to calculate the
display label.