Performance: reduce the detail types fetched during populate step
All details other than timestamp and status flags were moved out
of the main contacts table in qtcontacts-sqlite recently.
This means that it is relatively expensive to query detail types
such as gender/favorite/synctarget etc if they are not required.

Note that we still fetch the gender when populating the "all" list,
as we can use this data when determining valid merge targets.