[contacts] Add action plugin API and allow action invocation via SeasidePerson::triggerAction(). JB#55390
Add SeasideActionPlugin, which defines an interface for implementing
the actions that are triggered for a contact (call a contact, email a
contact etc.). DefaultSeasideActionPlugin provides a default
implementation that will launch the appropriate app for various
actions.

SeasideCache::actionPlugin() returns the default action plugin, which
is currently an instance of DefaultSeasideActionPlugin. Later on, we
can change this to find available action plugins through the Qt plugin
mechanism instead.

SeasidePerson::supportedActions() returns the available actions based
on those listed by the action plugin. SeasidePerson::triggerAction()
triggers an action using the action plugin.